Data on Scholarships in CAPES International Mobility Programs

The goal of capesData is to provide an attractive dataset for exploring and visualizing data on the International Mobility Programs funded by CAPES/Brazil. The CAPES database refers to international mobility programs for the period from 2010 to 2019.

Usage

capesData

Format

A data frame with 146036 rows and 14 columns:

ID_PROCESSO

Number of concession process.

NM_NIVEL

Type of scholarship.

AN_INICIO

Year in which the scholarship begins.

ME_INICIO

Month in which the scholarship begins.

AN_FIM

End year of the scholarship.

ME_FIM

End month of the scholarship.

NM_IES_ORIGEM_PRINCIPAL_DA

Name of the beneficiary's home institution.

NM_UF_IES_ORIGEM

Name of the Federation State of the beneficiary's institution of origin.

NM_IES_ESTUDO_PRINCIPAL_DA

Name of the beneficiary's study institution.

NM_PAIS_IES_ESTUDO

Country of destination of the beneficiary.

NM_GRANDE_AREA

Field of knowledge.

NM_AREA_CONHECIMENTO

Area of knowledge classified by CAPES.

NM_AREA_AVALIACAO

Area of project evaluation.

NM_PROGRAMA

Program Name.

Source

Created by Leonardo Biazoli with data provided by CAPES/BRAZIL https://dadosabertos.capes.gov.br/dataset/.

See also

Examples

capesData
#> # A tibble: 146,036 × 14
#>    ID_PROCESSO NM_NIVEL AN_INICIO ME_INICIO AN_FIM ME_FIM NM_IES_ORIGEM_PRINCI…¹
#>    <chr>       <chr>        <dbl> <chr>      <dbl> <chr>  <chr>                 
#>  1 3038-11-4   GRADUAÇ…      2011 08          2012 05     UNIVERSIDADE FEDERAL …
#>  2 3038-11-4   GRADUAÇ…      2011 08          2012 05     UNIVERSIDADE FEDERAL …
#>  3 402509      DOUTORA…      2007 04          2011 03     PONTIFÍCIA UNIVERSIDA…
#>  4 1091272     DOUTORA…      2012 03          2015 02     UNIVERSIDADE DE BRASÍ…
#>  5 623546      MESTRAD…      2009 12          2010 05     UNIVERSIDADE FEDERAL …
#>  6 0963-10-0   ESTÁGIO…      2010 05          2010 07     UNIVERSIDADE DE SÃO P…
#>  7 9904-11-5   DOUTORA…      2012 05          2013 04     UNIVERSIDADE FEDERAL …
#>  8 9904-11-5   DOUTORA…      2012 05          2013 04     UNIVERSIDADE FEDERAL …
#>  9 9202-12-9   DOUTORA…      2012 09          2013 04     UNIVERSIDADE DE BRASÍ…
#> 10 948932      DOUTORA…      2011 04          2011 09     UNIVERSIDADE FEDERAL …
#> # ℹ 146,026 more rows
#> # ℹ abbreviated name: ¹​NM_IES_ORIGEM_PRINCIPAL_DA
#> # ℹ 7 more variables: NM_UF_IES_ORIGEM <chr>, NM_IES_ESTUDO_PRINCIPAL_DA <chr>,
#> #   NM_PAIS_IES_ESTUDO <chr>, NM_GRANDE_AREA <chr>, NM_AREA_CONHECIMENTO <chr>,
#> #   NM_AREA_AVALIACAO <chr>, NM_PROGRAMA <chr>