Data on Scholarships in CAPES International Mobility Programs

The goal of capesData is to provide an attractive dataset for exploring and visualizing data on the International Mobility Programs funded by CAPES/Brazil. The CAPES database refers to international mobility programs for the period from 2010 to 2019.

Usage

capesData_raw

Format

A data frame with 146036 rows and 51 columns:

ID_PROCESSO

Number of concession process.

NM_NIVEL

Type of scholarship.

AN_INICIO

Year in which the scholarship begins.

ME_INICIO

Month in which the scholarship begins.

AN_FIM

End year of the scholarship.

ME_FIM

End month of the scholarship.

NM_IES_ORIGEM_PRINCIPAL_DA

Name of the beneficiary's home institution.

NM_UF_IES_ORIGEM

Name of the Federation State of the beneficiary's institution of origin.

NM_IES_ESTUDO_PRINCIPAL_DA

Name of the beneficiary's study institution.

NM_PAIS_IES_ESTUDO

Country of destination of the beneficiary.

NM_GRANDE_AREA

Field of knowledge.

NM_AREA_CONHECIMENTO

Area of knowledge classified by CAPES.

NM_AREA_AVALIACAO

Area of project evaluation.

NM_PROGRAMA

Program Name.

CD_MOEDA

Functional payment currency.

VL_TOTAL_RECEBIDO_MOEDA

Total amount received in functional currency.

VL_AUX_DESLOCAMENTO

Amount received as travel assistance.

VL_ADICIONAL_LOCALIDADE

Amount received as a locality bonus.

VL_ADICIONAL_DEPENDENTE

Additional amount paid to full doctoral scholarship holders with registered dependents.

VL_BOLSA

Amount received as a scholarship/fees.

VL_TAXA_ESCOLAR

Amount received for tuition and fees.

VL_DIF_AUX_DESLOCAMENTO

Amount received as differences/reimbursements for travel expenses that were performed.

VL_DIF_SEGURO_SAUDE

Amount received as differences/reimbursements for expenses performed with health insurance.

VL_SEGURO_SAUDE

Amount received as health insurance.

VL_TAXA_ESCOLAR_VALOR_P

Amount received for small tuition and fees.

VL_OUTROS_CREDITOS

Amount received for the payment of extraordinary expenses.

VL_DIF_MENSALIDADE

Amount received as differences/reimbursements for expenses performed with tuition.

VL_OUTROS_DEBITOS

Amount received as differences/refunds for other debts.

VL_DIF_DEPENDENTES

Amount received as differences/reimbursements for expenses performed with additional dependent.

VL_MENSALIDADE

Amount received as a stipend.

VL_AUX_INSTALACAO

Amount received as installation aid.

VL_AUX_SEGURO_SAUDE_ANUAL

Amount received in annual installments as health insurance.

VL_MENSALIDADES_AGENDADAS

Amount received as scheduled tuition.

VL_REEMBOLSO_TAXA_ESCOLAR

Amount received for the refund of tuition and fees paid by the beneficiary.

VL_AUX_MORADIA

Amount received as housing allowance.

VL_DIF_AUX_INSTALACAO

Amount received as differences/refunds for installation aid.

VL_REEMBOLSO_PASSAGEM_AEREA

Amount received as differences/refunds for airfare.

VL_AUX_MATERIAL_DIDATICO

Amount received as aid for teaching materials, paid under the Ciencia Sem Fronteiras Program.

VL_AUX_SEGURO_SAUDE

Amount received as health insurance benefit.

VL_AJUDA_CUSTO_CAPES_SETEC

Amount received as a subsistence allowance, paid under the CAPES/SETEC program.

VL_MENSALIDADE_DEMANDA

Amount received as a monthly fee on demand.

VL_AUX_DESLOCAMENTO_DEMANDA

Amount received as travel allowance on demand.

VL_MENSALIDADE_LIC_MATERNIDADE

Amount received as tuition during maternity leave by the scholarship holder during the duration of the process.

VL_ESCOLA_ALTOS_ESTUDOS

Amount received as a scholarship under the Escola de Altos Estudos program.

VL_AJUDA_CUSTO

Amount received as a subsistence allowance.

VL_AUX_INSTALACAO_DEPENDENTE

Additional amount received as installation aid, paid to full doctoral scholarship holders with registered dependents.

VL_AUX_SEGURO_SAUDE_DEPENDENTE

Additional amount received as health insurance aid, paid to full doctoral scholarship holders with registered dependents.

VL_AJUDA_CUSTO_MTUR

Amount received as a subsistence allowance, paid under the CAPES/MTUR program.

VL_AUX_DESLOCAMENTO_RETORNO_D

Amount received as travel allowance for return.

VL_AUX_SEGURO_SAUDE_DEMANDA

Amount received as health insurance on demand.

VL_AUX_DESLOCAMENTO_PESQUISA

Amount received as travel allowance for research.

Source

Created by Leonardo Biazoli with data provided by CAPES/BRAZIL https://dadosabertos.capes.gov.br/dataset/.

See also

Examples

capesData_raw
#> # A tibble: 146,036 × 51
#>    ID_PROCESSO NM_NIVEL AN_INICIO ME_INICIO AN_FIM ME_FIM NM_IES_ORIGEM_PRINCI…¹
#>    <chr>       <chr>        <dbl> <chr>      <dbl> <chr>  <chr>                 
#>  1 3038-11-4   GRADUAÇ…      2011 08          2012 05     UNIVERSIDADE FEDERAL …
#>  2 3038-11-4   GRADUAÇ…      2011 08          2012 05     UNIVERSIDADE FEDERAL …
#>  3 402509      DOUTORA…      2007 04          2011 03     PONTIFÍCIA UNIVERSIDA…
#>  4 1091272     DOUTORA…      2012 03          2015 02     UNIVERSIDADE DE BRASÍ…
#>  5 623546      MESTRAD…      2009 12          2010 05     UNIVERSIDADE FEDERAL …
#>  6 0963-10-0   ESTÁGIO…      2010 05          2010 07     UNIVERSIDADE DE SÃO P…
#>  7 9904-11-5   DOUTORA…      2012 05          2013 04     UNIVERSIDADE FEDERAL …
#>  8 9904-11-5   DOUTORA…      2012 05          2013 04     UNIVERSIDADE FEDERAL …
#>  9 9202-12-9   DOUTORA…      2012 09          2013 04     UNIVERSIDADE DE BRASÍ…
#> 10 948932      DOUTORA…      2011 04          2011 09     UNIVERSIDADE FEDERAL …
#> # ℹ 146,026 more rows
#> # ℹ abbreviated name: ¹​NM_IES_ORIGEM_PRINCIPAL_DA
#> # ℹ 44 more variables: NM_UF_IES_ORIGEM <chr>,
#> #   NM_IES_ESTUDO_PRINCIPAL_DA <chr>, NM_PAIS_IES_ESTUDO <chr>,
#> #   NM_GRANDE_AREA <chr>, NM_AREA_CONHECIMENTO <chr>, NM_AREA_AVALIACAO <chr>,
#> #   NM_PROGRAMA <chr>, CD_MOEDA <chr>, VL_TOTAL_RECEBIDO_MOEDA <dbl>,
#> #   VL_AUX_DESLOCAMENTO <dbl>, VL_ADICIONAL_LOCALIDADE <dbl>, …